+++ core/src/main/java/org/web3j/tx/gas/ContractEIP1559GasProvider.java
@@ -15,11 +15,9 @@ package org.web3j.tx.gas;
 import java.math.BigInteger;
 
 public interface ContractEIP1559GasProvider extends ContractGasProvider {
-    boolean isEIP1559Enabled();
-
     long getChainId();
 
-    BigInteger getMaxFeePerGas(String contractFunc);
+    BigInteger getMaxFeePerGas();
 
-    BigInteger getMaxPriorityFeePerGas(String contractFunc);
+    BigInteger getMaxPriorityFeePerGas();
 }